@@ -4,12 +4,14 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@datadog/browser-rum": "^6.22.0",
+    "@datadog/browser-rum-react": "^6.22.0",
     "@testing-library/jest-dom": "^5.15.0",
-    "@testing-library/react": "^11.2.7",
+    "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^12.8.3",
     "dotenv": "^10.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-router-dom": "^6.0.2",
     "web-vitals": "^1.1.2"
   },
