@@ -1,14 +1,11 @@
-// import logo from './logo.svg';
+import { createBrowserRouter } from '@datadog/browser-rum-react/react-router-v6';
 import Header from './components/main/Header.jsx';
 import Home from './components/Home.jsx';
 import Products from './components/Products.jsx';
 import Services from './components/Services.jsx';
 import About from './components/About.jsx';
 import Contact from './components/Contact.jsx';
-import {
-  Routes,
-  Route
-} from "react-router-dom";
+import { Outlet } from "react-router-dom";
 import Footer from './components/main/Footer.jsx';
 import { useState } from 'react';
 
@@ -21,13 +18,7 @@ function MainWrapper() {
       <Header dropMenu={dropMenu} setDropMenu={setDropMenu} />
 
       <div id="Content">
-        <Routes>
-          <Route path={"/"} exact element={<Home setDropMenu={setDropMenu} />} />
-          <Route path={"/products/*"} element={<Products setDropMenu={setDropMenu} />} />
-          <Route path={"/services"} element={<Services setDropMenu={setDropMenu} />} />
-          <Route path={"/about"} element={<About setDropMenu={setDropMenu} />} />
-          <Route path={"/contact"} element={<Contact setDropMenu={setDropMenu} />} />
-        </Routes>
+        <Outlet />
       </div>
       <Footer />
     </div>
@@ -35,3 +26,33 @@ function MainWrapper() {
 }
 
 export default MainWrapper;
+
+// Create router configuration with Datadog's router
+export const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <MainWrapper />,
+    children: [
+      {
+        path: '/',
+        element: <Home setDropMenu={(val) => val} />
+      },
+      {
+        path: '/products/*',
+        element: <Products setDropMenu={(val) => val} />
+      },
+      {
+        path: '/services',
+        element: <Services setDropMenu={(val) => val} />
+      },
+      {
+        path: '/about',
+        element: <About setDropMenu={(val) => val} />
+      },
+      {
+        path: '/contact',
+        element: <Contact setDropMenu={(val) => val} />
+      }
+    ]
+  }
+]);
